pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["synthetic_data", "attack_scenarios", "mip"]

[project]
name = "ukraine-medical-logistics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "contextily>=1.7.0",
    "folium>=0.20.0",
    "geopandas>=1.1.3",
    "gurobipy>=13.0.1",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "osmnx>=2.1.0",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.8.0",
    "shapely>=2.1.2",
]
